replace nvim config with LazyVim
