+++ dot_config/nvim/README.md
@@ -1,8 +1,45 @@
-# Neovim Config
+# LazyVim
 
-Most of this is ripped from [Neovim-from-scratch](https://github.com/LunarVim/Neovim-from-scratch).
+A starter template for [LazyVim](https://github.com/LazyVim/LazyVim)
 
-And some other resources:
-- [Primeagen .dotfiles](https://github.com/ThePrimeagen/.dotfiles)
-- [TJ Devries dots](https://github.com/tjdevries/config_manager)
-- [bashbunni dotfiles](https://github.com/bashbunni/dotfiles)
+## 🚀 Getting Started
+
+This repo contains an example setup for
+[LazyVim](https://github.com/LazyVim/LazyVim)
+
+### 1. Make a backup of your current Neovim files:
+
+```sh
+mv ~/.config/nvim ~/.config/nvim.bak
+mv ~/.local/share/nvim ~/.local/share/nvim.bak
+```
+
+### 2. Clone the starter
+
+```sh
+git clone https://github.com/LazyVim/starter ~/.config/nvim
+```
+
+### 3. Start Neovim!
+
+```sh
+nvim
+```
+
+Refer to the comments in the files on how to customize **LazyVim**.
+
+## 📂 File Structure
+
+<pre>
+~/.config/nvim
+├── lua
+│   ├── config
+│   │   ├── autocmds.lua
+│   │   ├── keymaps.lua
+│   │   ├── lazy.lua
+│   │   └── options.lua
+│   └── plugins
+│       └── example.lua
+├── init.lua
+└── stylua.toml
+</pre>
