+++ dot_config/nvim/dot_gitignore
@@ -1,0 +1,8 @@
+tt.*
+.tests
+doc/tags
+debug
+.repro
+foo.*
+*.log
+data
