+++ dot_config/nvim/init.lua
@@ -1,18 +1,2 @@
-vim.cmd("source ~/.vimrc")
-
-require("user.options")
-require("user.plugins")
-require("user.autocommands")
-require("user.impatient")
-require("user.autopairs")
-require("user.colorscheme")
-require("user.lualine")
-require("user.cmp")
-require("user.toggleterm")
-require("user.lsp")
-require("user.treesitter")
-require("user.telescope")
-require("user.gitsigns")
-require("user.indent_blankline")
-require("user.whichkey")
-require("user.vimwiki")
+-- bootstrap lazy.nvim, LazyVim and your plugins
+require("config.lazy")
