+++ dot_config/nvim/lua/plugins/cmp.lua
@@ -1,0 +1,8 @@
+return {
+  {
+    "L3MON4D3/LuaSnip",
+    config = function()
+      require("luasnip.loaders.from_vscode").lazy_load({ paths = "~/.config/nvim/snippets" })
+    end,
+  },
+}
