+++ dot_config/nvim/lua/plugins/core.lua
@@ -1,0 +1,19 @@
+return {
+  {
+    "LazyVim/LazyVim",
+    opts = {
+      colorscheme = "tokyonight-storm",
+    },
+  },
+
+  { "akinsho/bufferline.nvim", enabled = false },
+  { "ggandor/leap.nvim", enabled = false },
+  { "ggandor/flit.nvim", enabled = false },
+
+  -- extra colorschemes
+  "mangeshrex/everblush.vim",
+  "doums/darcula",
+  "nyoom-engineering/oxocarbon.nvim",
+  "rose-pine/neovim",
+  "morhetz/gruvbox",
+}
