+++ dot_config/nvim/lua/plugins/toggleterm.lua
@@ -1,0 +1,24 @@
+return {
+  {
+    "akinsho/toggleterm.nvim",
+    config = function()
+      require("toggleterm").setup({
+        size = 15,
+        open_mapping = [[<c-\>]],
+        direction = "horizontal",
+        close_on_exit = true, -- close the terminal window when the process exits
+        shell = vim.o.shell, -- change the default shell
+        -- This field is only relevant if direction is set to 'float'
+        float_opts = {
+          border = "single",
+          width = 80,
+          height = 20,
+          highlights = {
+            border = "Normal",
+            background = "Normal",
+          },
+        },
+      })
+    end,
+  },
+}
