--- dot_config/nvim/lua/user/autocommands.lua
@@ -1,10 +1,0 @@
-vim.cmd [[
-  augroup _general_settings
-    autocmd!
-    autocmd FileType qf,help,man,lspinfo nnoremap <silent> <buffer> q :close<CR> 
-    autocmd TextYankPost * silent!lua require('vim.highlight').on_yank({higroup = 'Visual', timeout = 200}) 
-    autocmd BufWinEnter * :set formatoptions-=cro
-    autocmd FileType qf set nobuflisted
-    autocmd BufWritePre * lua vim.lsp.buf.format()
-  augroup end
-]]
