--- dot_config/nvim/lua/user/colorscheme.lua
@@ -1,5 +1,0 @@
-vim.g.nord_italic = false
-vim.g.nord_bold = false
-
--- Load the colorscheme
-vim.cmd("colorscheme tokyonight-storm")
