--- dot_config/nvim/lua/user/impatient.lua
@@ -1,6 +1,0 @@
-local status_ok, impatient = pcall(require, "impatient")
-if not status_ok then
-  return
-end
-
-impatient.enable_profile()
