--- dot_config/nvim/lua/user/indent_blankline.lua
@@ -1,6 +1,0 @@
-local status_ok, indent_blankline = pcall(require, "indent_blankline")
-if not status_ok then
-	return
-end
-
-indent_blankline.setup({ filetype = { "yaml" } })
