--- dot_config/nvim/lua/user/lsp/null-ls.lua
@@ -1,28 +1,0 @@
-local null_ls_status_ok, null_ls = pcall(require, "null-ls")
-if not null_ls_status_ok then
-	return
-end
-
--- https://github.com/jose-elias-alvarez/null-ls.nvim/tree/main/lua/null-ls/builtins/formatting
-local formatting = null_ls.builtins.formatting
--- https://github.com/jose-elias-alvarez/null-ls.nvim/tree/main/lua/null-ls/builtins/diagnostics
-local diagnostics = null_ls.builtins.diagnostics
-
-null_ls.setup({
-	debug = false,
-	sources = {
-		formatting.black.with({ extra_args = { "--fast" } }),
-		formatting.isort,
-		formatting.brittany,
-		formatting.mix,
-		formatting.stylua,
-		formatting.prettier,
-    formatting.scalafmt,
-    formatting.shfmt,
-    formatting.sqlfluff,
-    formatting.yamlfmt,
-    diagnostics.hadolint,
-    diagnostics.yamllint,
-    -- diagnostics.mypy,
-	},
-})
