--- dot_config/nvim/lua/user/lualine.lua
@@ -1,8 +1,0 @@
-local status_ok, lualine = pcall(require, "lualine")
-if not status_ok then
-	return
-end
-
-lualine.setup({
-	options = { theme = "tokyonight" },
-})
