+++ dot_config/nvim/snippets/snippets/flux.json
