+++ dot_config/nvim/stylua.toml
@@ -1,0 +1,3 @@
+indent_type = "Spaces"
+indent_width = 2
+column_width = 120
